+++ server/src/migrations/202104130959-create-table-cars.ts
@@ -55,7 +55,7 @@ export = {
         allowNull: false
       },
       engineSize: {
-        type: DataTypes.SMALLINT,
+        type: DataTypes.FLOAT,
         allowNull: false
       },
       fuelType: {
